% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{akaike_weights}
\alias{akaike_weights}
\title{Calculate Akaike Weights for a set of models.}
\usage{
akaike_weights(icV)
}
\arguments{
\item{icV}{Vector of information criteria.}
}
\value{
Vector of Akaike weights
}
\description{
Calculate Akaike Weights for a set of models.
}
