% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_drop.R
\name{bind_rows_drop}
\alias{bind_rows_drop}
\title{Bind the rows of a list of data frames while dropping all columns not in
common}
\usage{
bind_rows_drop(l)
}
\arguments{
\item{l}{A list of dataframes}
}
\value{
A single dataframe that has all the rows from the original list of
data frames, but only the columns that are represented in all data frames.
Note that no identifier is provided from where the rows came from at present
}
\description{
Bind the rows of a list of data frames while dropping all columns not in
common
}
