% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clamp_complex.R
\name{clamp_complex}
\alias{clamp_complex}
\title{Clamp complex numbers}
\usage{
clamp_complex(
  x,
  lower = complex(real = -Inf, imaginary = -Inf),
  upper = complex(real = Inf, imaginary = Inf)
)
}
\arguments{
\item{x}{A vector of complex numbers with one or more elements.}

\item{lower}{The lower bound.}

\item{upper}{The upper bound.}
}
\value{
A vector of complex numbers.
}
\description{
Clamp the real and imaginary parts of complex numbers to values between the respective lower and upper bounds.
}
\examples{
clamp_complex(3 + 3i, 1 + 4i, 2 + 5i)
}
