% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_by_funs.R
\name{filter_by_funs}
\alias{filter_by_funs}
\title{Filter a data frame accoring to descriptive functions}
\usage{
filter_by_funs(data, variable, .funs)
}
\arguments{
\item{data}{The data frame that includes the variables to compute
descriptive statistics on.}

\item{variable}{The variable to filter on.}

\item{.funs}{The functions to use in filtering.}
}
\value{
A tidy data frame with the requested descriptive statistics
for the requested variables
}
\description{
This function allows you to subset a data frame by any set
of descriptive statistics. For example, you could extract rows
that are eqaul to specific quantiles, or the min and the max.
This is a dplyr-like function, and works well with functions
such as \code{\link[dplyr]{group_by}}.
}
\examples{

library(dplyr)
storms \%>\%
 filter_by_funs(wind, funs(min,
                           quantile(., .25, na.rm = TRUE),
                           quantile(., .50, na.rm = TRUE),
                           quantile(., .75, na.rm = TRUE),
                           max))
storms \%>\%
  group_by(year) \%>\%
  filter_by_funs(wind, funs(min,max)) \%>\%
  arrange(year)
}
