% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descrips.R
\name{gather_spread_descrips}
\alias{gather_spread_descrips}
\title{Formatting for descrips}
\usage{
gather_spread_descrips(smry, grouping, order)
}
\arguments{
\item{smry}{The smry output from \code{\link{descrips}}}

\item{grouping}{String stating the grouping factor, and NULL
otherwise.}

\item{order}{The order the columns should be displayed in the
final output}
}
\description{
Internal function to format the descrips in a more eye-ball
friendly format
}
