% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nms.R
\name{nms}
\alias{nms}
\title{Easily see variable index locations}
\usage{
nms(d)
}
\arguments{
\item{d}{Data frame from which to report variable names and their locations}
}
\value{
Matrix of the names from the dataset and their location
}
\description{
Function returns the variable names and their location in a matrix
}

