% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_residual_vcov.R
\name{pull_residual_vcov}
\alias{pull_residual_vcov}
\title{Pull the residual variance-covariance matrix from an lme4::lmer model}
\usage{
pull_residual_vcov(model)
}
\arguments{
\item{model}{A fitted lme4::lmer model}
}
\value{
The residual variance covariance matrix (which is a sparse matrix)
}
\description{
This function wraps the code supplied by Ben Bolker from StackOverflow
\link{https://stackoverflow.com/a/45655597}{here}.
}
