% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/regPlot.R
\name{regPlot}
\alias{regPlot}
\title{Scatterplot of any two variables, as well as the regression line with the
correlation reported. Alternatively, a regression model can be fed to the
function with an equivalent plot reported.}
\usage{
regPlot(x, y, lcol = "blue", se = TRUE, secol = c(0, 0, 0.5, 0.2), ...)
}
\arguments{
\item{x}{Predictor variable (plotted on the x-axis)}

\item{y}{Outcome (plotted on the y-axis)}

\item{lcol}{Color of the regression line. Defaults to blue.}

\item{se}{Logical, defaults to \code{TRUE}. Should the standard error
around the regression line be plotted?}

\item{secol}{Color and transparency of the standard error shading. Vector
of four elements as follows: red, green, blue, alpha, where alpha is the
transparency parameter (higher values increase the opacity).}

\item{mod}{Fitted regression model from \code{\link[stats]{lm}}. Alternative
to supplying \code{x} and \code{y}.}
}
\value{
Scatterplot of the relation between two variables (linear
regression)
}
\description{
Scatterplot of any two variables, as well as the regression line with the
correlation reported. Alternatively, a regression model can be fed to the
function with an equivalent plot reported.
}

