% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_empty_rows.R
\name{rm_empty_rows}
\alias{rm_empty_rows}
\title{Remove rows with empty data}
\usage{
rm_empty_rows(d, ...)
}
\arguments{
\item{d}{The data frame.}

\item{...}{Optional columns to include. If included, rows will only be
removed if the data are missing across all columns included. If \code{NULL}
only rows with complete missing data across all columns will be removed
(i.e., equivalent to \link[janitor]{remove_empty_rows})}
}
\description{
This function works similarly to \link[janitor]{remove_empty_rows} function
from the janitor package, except that you can pass the list of columns to
remove missing data from. You can then easily eliminate any rows that have
complete missing data across any generic set of columns
}
