% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.R
\name{se}
\alias{se}
\title{Calculate the se of the mean}
\usage{
se(v)
}
\arguments{
\item{v}{Vector to calculate the se of the mean on.}
}
\description{
This function is pretty simple. Just calculates the standard error of the
mean for a generic vector. For some odd reason this is not included in the 
base \emph{stats} package, so I put it in mine. Removes missing data by
default
}

