% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sundry.R
\docType{package}
\name{sundry}
\alias{sundry}
\title{sundry: A sundry of convenience functions}
\description{
This pacakge contains miscellaneous convenience functions to make
data analysis and manipulation more efficient.
}
