% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/weights.R
\name{weights}
\alias{weights}
\title{Calculate Akaike Weights for a set of models.}
\usage{
weights(icV)
}
\arguments{
\item{icV}{Vector of information criteria.}
}
\value{
Vector of Akaike weights
}
\description{
Calculate Akaike Weights for a set of models.
}

