% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.mplus.R
\name{write.mplus}
\alias{write.mplus}
\title{Write .dat data file for Mplus. Note variable names are printed to the
console.}
\usage{
write.mplus(d, fileName, dir = getwd(), na.string = "999", ...)
}
\arguments{
\item{d}{Data frame to  write out.}

\item{fileName}{Name of the file to be written out. Note that the extension
is added automatically, and only the file name should be supplied.}

\item{dir}{Directory to write the file. Defaults to the current working
directory.}

\item{na.string}{String to fill in missing values. Defaults to "999"}

\item{...}{Additional arguments passed to \code{\link[utils]{write.table}}}
}
\value{
Mplus .dat file written to the current working directory, with
 variable names printed to the console (for easy copy and paste into a .inp
 Mplus file)
}
\description{
Write .dat data file for Mplus. Note variable names are printed to the
console.
}
