% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{aic_weights}
\alias{aic_weights}
\title{Calculate Akaike Weights for a set of models.}
\usage{
aic_weights(...)
}
\arguments{
\item{...}{The candidate models to compare}
}
\value{
Vector of Akaike weights
}
\description{
Calculate Akaike Weights for a set of models.
}
