% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nms.R
\name{nms}
\alias{nms}
\title{Easily see variable index locations}
\usage{
nms(d)
}
\arguments{
\item{d}{Data frame from which to report variable names and their locations}
}
\value{
data frame of the names from the dataset and their location as the
row names
}
\description{
Function returns the variable names and their location in a data frame (the
row names index their locations)
}
