% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readFiles.R
\name{readFiles}
\alias{readFiles}
\title{Batch read csv files, return all files in a list}
\usage{
readFiles(dir = getwd(), pat = NULL, ...)
}
\arguments{
\item{dir}{Directory where csv files are stored. Defaults to the current
working directory}

\item{pat}{Optional string pattern of files in the directory}

\item{...}{Additonal arguments passed to \code{\link[utils]{read.csv}}.}
}
\description{
Batch read csv files, return all files in a list
}

